Add initial auto network joining
Added the initial pieces for auto network joining. Currently, there's no
way to stop it. I'm also unsure if I want to have this autojoining
program running in the Traefik container alongside it, or if I want it
to be a daemon on the local machine, started up by `falcon up`. I'm
leaning towards daemonization, as it follows the spirit of
containerization better, and it might allow for some cool learning
opportunities about Go and daemonizing.