Update Syncing Strategy
Recently, I made a great deal of progress on the syncing over in the
main falcon repo. However, I decided that I would rather run the network
auto-joining code inside the proxy container itself, rather than deal
with the many problems that could arise from making the auto-joining
code into a system daemon. Instead, we can let Docker handle all that
for us!

Additionally, the syncing strategy has been updated. Instead of joining
or leaving based on a network event, we simply wait for network events
to let us know when to start computing any actions we should take. Then,
we look at the list of all valid networks and all networks the proxy
is a part of, and then decide what actions, if any, we need to take.