Switch to config directory for dynamic config
In order to fix some issues with docker not notifying Traefik that the
dynamic config has changed, we're doing a quick path change to where the
dynamic config is mounted inside the container. In addition, fixed some
comments that were either outdated or had some typos and ran gofmt on
all files.