Fix entrypoint issue
I decided to use the array style for calling CMD in the dockerfile,
which doesn't allocate a shell, but I forgot that I'm literally
executing a shell file, so a) I don't need to do that and b) it's nice
to let the shell expand the WORKDIR environment variable rather than
hardcode it. This commit changes switches back to the plain ol string
CMD methodology.